Struct futures_mio::Loop

pub struct Loop {
    // some fields omitted
}

An event loop.

The event loop is the main source of blocking in an application which drives all other I/O events and notifications happening. Each event loop can have multiple handles pointing to it, each of which can then be used to create various I/O objects to interact with the event loop in interesting ways.

Methods

impl Loop

fn new() -> Result<Loop>

Creates a new event loop, returning any error that happened during the creation.

fn handle(&self) -> LoopHandle

Generates a handle to this event loop used to construct I/O objects and send messages.

Handles to an event loop are cloneable as well and clones will always refer to the same event loop.

fn pin(&self) -> LoopPin

Returns a "pin" of this event loop which cannot be sent across threads but can be used as a proxy to the event loop itself.

Currently the primary use for this is to use as a handle to add data to the event loop directly. The LoopPin::add_loop_data method can be used to immediately create instances of LoopData structures.

fn add_loop_data<A>(&self, a: A) -> LoopData<A> where A: 'static

Creates a new LoopData<A> handle by associating data to be directly stored by this event loop.

This function is useful for when storing non-Send data inside of a future. The LoopData<A> handle is itself Send + 'static regardless of the underlying A. That is, for example, you can create a handle to some data that contains an Rc, for example.

fn run<F: Future>(&mut self, f: F) -> Result<F::Item, F::Error>

Runs a future until completion, driving the event loop while we're otherwise waiting for the future to complete.

Returns the value that the future resolves to.