Struct futures_mio::LoopData

pub struct LoopData<A: Any> {
    // some fields omitted
}

A handle to data that is owned by an event loop thread, and is only accessible on that thread itself.

This structure is created by the LoopHandle::add_loop_data method which will return a future resolving to one of these references. A LoopData<A> handle is Send regardless of what A is, but the internal data can only be accessed on the event loop thread itself.

Internally this reference also stores a handle to the event loop that the data originated on, so it knows how to go back to the event loop to access the data itself.

Methods

impl<A: Any> LoopData<A>

fn get(&self) -> Option<&A>

Gets a shared reference to the underlying data in this handle.

Returns None if it is not called from the event loop thread that this LoopData<A> is associated with, or Some with a reference to the data if we are indeed on the event loop thread.

fn get_mut(&mut self) -> Option<&mut A>

Gets a mutable reference to the underlying data in this handle.

Returns None if it is not called from the event loop thread that this LoopData<A> is associated with, or Some with a reference to the data if we are indeed on the event loop thread.

fn executor(&self) -> Arc<Executor>

Acquire the executor associated with the thread that owns this LoopData<A>'s data.

If the get and get_mut functions above return None, then this data is being polled on the wrong thread to access the data, and to make progress a future may need to migrate to the actual thread which owns the relevant data.

This executor can in turn be passed to Task::poll_on, which will then move the entire future to be polled on the right thread.