Struct futures_mio::LoopHandle

pub struct LoopHandle {
    // some fields omitted
}

Handle to an event loop, used to construct I/O objects, send messages, and otherwise interact indirectly with the event loop itself.

Handles can be cloned, and when cloned they will still refer to the same underlying event loop.

Methods

impl LoopHandle

fn add_source(&self, source: IoSource) -> AddSource

Add a new source to an event loop, returning a future which will resolve to the token that can be used to identify this source.

When a new I/O object is created it needs to be communicated to the event loop to ensure that it's registered and ready to receive notifications. The event loop with then respond with a unique token that this handle can be identified with (the resolved value of the returned future).

This token is then passed in turn to each of the methods below to interact with notifications on the I/O object itself.

Panics

The returned future will panic if the event loop this handle is associated with has gone away, or if there is an error communicating with the event loop.

fn schedule(&self, tok: usize, task: &mut Task)

Begin listening for events on an event loop.

Once an I/O object has been registered with the event loop through the add_source method, this method can be used with the assigned token to begin awaiting notifications.

The dir argument indicates how the I/O object is expected to be awaited on (either readable or writable) and the wake callback will be invoked. Note that one the wake callback is invoked once it will not be invoked again, it must be re-scheduled to continue receiving notifications.

Panics

This function will panic if the event loop this handle is associated with has gone away, or if there is an error communicating with the event loop.

fn deschedule(&self, tok: usize)

Stop listening for events on an event loop.

Once a callback has been scheduled with the schedule method, it can be unregistered from the event loop with this method. This method does not guarantee that the callback will not be invoked if it hasn't already, but a best effort will be made to ensure it is not called.

Panics

This function will panic if the event loop this handle is associated with has gone away, or if there is an error communicating with the event loop.

fn drop_source(&self, tok: usize)

Unregister all information associated with a token on an event loop, deallocating all internal resources assigned to the given token.

This method should be called whenever a source of events is being destroyed. This will ensure that the event loop can reuse tok for another I/O object if necessary and also remove it from any poll notifications and callbacks.

Note that wake callbacks may still be invoked after this method is called as it may take some time for the message to drop a source to reach the event loop. Despite this fact, this method will attempt to ensure that the callbacks are not invoked, so pending scheduled callbacks cannot be relied upon to get called.

Panics

This function will panic if the event loop this handle is associated with has gone away, or if there is an error communicating with the event loop.

fn add_timeout(&self, at: Instant) -> AddTimeout

Adds a new timeout to get fired at the specified instant, notifying the specified task.

fn update_timeout(&self, timeout: &TimeoutToken, task: &mut Task)

Updates a previously added timeout to notify a new task instead.

Panics

This method will panic if the timeout specified was not created by this loop handle's add_timeout method.

fn cancel_timeout(&self, timeout: &TimeoutToken)

Cancel a previously added timeout.

Panics

This method will panic if the timeout specified was not created by this loop handle's add_timeout method.

fn add_loop_data<F, A>(&self, f: F) -> AddLoopData<F, A> where F: FnOnce() -> A + Send + 'static, A: 'static

Schedules a closure to add some data to event loop thread itself.

This function is useful for when storing non-Send data inside of a future. This returns a future which will resolve to a LoopData<A> handle, which is itself Send + 'static regardless of the underlying A. That is, for example, you can create a handle to some data that contains an Rc, for example.

This function takes a closure which may be sent to the event loop to generate an instance of type A. The closure itself is required to be Send + 'static, but the data it produces is only required to adhere to 'static.

If the returned future is polled on the event loop thread itself it will very cheaply resolve to a handle to the data, but if it's not polled on the event loop then it will send a message to the event loop to run the closure f, generate a handle, and then the future will yield it back.

fn shutdown(&self)

Send a message to the associated event loop that it should shut down, or otherwise break out of its current loop of iteration.

This method does not forcibly cause the event loop to shut down or perform an interrupt on whatever task is currently running, instead a message is simply enqueued to at a later date process the request to stop looping ASAP.

Panics

This function will panic if the event loop this handle is associated with has gone away, or if there is an error communicating with the event loop.

impl LoopHandle

fn tcp_listen(self, addr: &SocketAddr) -> IoFuture<TcpListener>

Create a new TCP listener associated with this event loop.

The TCP listener will bind to the provided addr address, if available, and will be returned as a future. The returned future, if resolved successfully, can then be used to accept incoming connections.

fn tcp_connect(self, addr: &SocketAddr) -> IoFuture<TcpStream>

Create a new TCP stream connected to the specified address.

This function will create a new TCP socket and attempt to connect it to the addr provided. The returned future will be resolved once the stream has successfully connected. If an error happens during the connection or during the socket creation, that error will be returned to the future instead.

impl LoopHandle

fn udp_bind(self, addr: &SocketAddr) -> IoFuture<UdpSocket>

Create a new UDP socket bound to the specified address.

This function will create a new UDP socket and attempt to bind it to the addr provided. The returned future will be resolved once the socket has successfully bound. If an error happens during the binding or during the socket creation, that error will be returned to the future instead.

impl LoopHandle

fn timeout(self, dur: Duration) -> IoFuture<Timeout>

Creates a new timeout which will fire at dur time into the future.

This function will return a future that will resolve to the actual timeout object. The timeout object itself is then a future which will be set to fire at the specified point in the future.

fn timeout_at(self, at: Instant) -> IoFuture<Timeout>

Creates a new timeout which will fire at the time specified by at.

This function will return a future that will resolve to the actual timeout object. The timeout object itself is then a future which will be set to fire at the specified point in the future.