Struct futures_mio::Receiver

pub struct Receiver<T> {
    // some fields omitted
}

The receiving half of a channel used for processing messages sent by a Sender.

A Receiver cannot be cloned and is not Sync, so only one thread can receive messages at a time.

This type is created by the LoopHandle::channel method.

Methods

impl<T> Receiver<T>

fn recv(&self) -> Poll<T, Error>

Attempts to receive a message sent on this channel.

This method will attempt to dequeue any messages sent on this channel from any corresponding sender. If no message is available, but senders are still detected, then Poll::NotReady is returned and the current future task is scheduled to receive a notification when a message is available.

If an I/O error happens or if all senders have gone away (the channel is disconnected) then Poll::Err will be returned.