Struct futures_mio::Source

pub struct Source<E: ?Sized> {
    // some fields omitted
}

Type of I/O objects inserted into the event loop, created by Source::new.

Methods

impl<E> Source<E>

fn new(e: E) -> Source<E>

Creates a new Source wrapping the provided source of events.

impl<E: ?Sized> Source<E>

fn take_readiness(&self) -> Option<Ready>

Consumes the last readiness notification that this source received.

Currently sources receive readiness notifications on an edge-basis. That is, once you receive a notification that an object can be read, you won't receive any more notifications until all of that data has been read.

The event loop will fill in this information and then inform futures that they're ready to go with the schedule method, and then the poll method can use this to figure out what happened.

fn io(&self) -> &E

Gets access to the underlying I/O object.