Struct futures_mio::TcpListener

pub struct TcpListener {
    // some fields omitted
}

An I/O object representing a TCP socket listening for incoming connections.

This object can be converted into a stream of incoming connections for various forms of processing.

Methods

impl TcpListener

fn from_listener(listener: TcpListener, addr: &SocketAddr, handle: LoopHandle) -> IoFuture<TcpListener>

Create a new TCP listener from the standard library's TCP listener.

This method can be used when the LoopHandle::tcp_listen method isn't sufficient because perhaps some more configuration is needed in terms of before the calls to bind and listen.

This API is typically paired with the net2 crate and the TcpBuilder type to build up and customize a listener before it's shipped off to the backing event loop. This allows configuration of options like SO_REUSEPORT, binding to multiple addresses, etc.

The addr argument here is one of the addresses that listener is bound to and the listener will only be guaranteed to accept connections of the same address type currently.

Finally, the handle argument is the event loop that this listener will be bound to.

The platform specific behavior of this function looks like:

On Unix, the socket is placed into nonblocking mode and connections can be accepted as normal

On Windows, the address is stored internally and all future accepts will only be for the same IP version as addr specified. That is, if addr is an IPv4 address then all sockets accepted will be IPv4 as well (same for IPv6).

fn poll_read(&self) -> Poll<(), Error>

Test whether this socket is ready to be read or not.

fn local_addr(&self) -> Result<SocketAddr>

Returns the local address that this listener is bound to.

This can be useful, for example, when binding to port 0 to figure out which port was actually bound.

fn incoming(self) -> IoStream<(TcpStream, SocketAddr)>

Consumes this listener, returning a stream of the sockets this listener accepts.

This method returns an implementation of the Stream trait which resolves to the sockets the are accepted on this listener.