Struct futures_mio::timer_wheel::TimerWheel

pub struct TimerWheel<T> {
    // some fields omitted
}

An implementation of a timer wheel where data can be associated with each timer firing.

This structure implements a timer wheel data structure where each timeout has a piece of associated data, T. A timer wheel supports O(1) insertion and removal of timers, as well as quickly figuring out what needs to get fired.

Note, though, that the resolution of a timer wheel means that timeouts will not arrive promptly when they expire, but rather in certain increments of each time. The time delta between each slot of a time wheel is of a fixed length, meaning that if a timeout is scheduled between two slots it'll end up getting scheduled into the later slot.

Methods

impl<T> TimerWheel<T>

fn new() -> TimerWheel<T>

Creates a new timer wheel configured with no timeouts and with the default parameters.

Currently this is a timer wheel of length 256 with a 100ms time resolution.

fn insert(&mut self, at: Instant, data: T) -> Timeout

Creates a new timeout to get fired at a particular point in the future.

The timeout will be associated with the specified data, and this data will be returned from poll when it's ready.

The returned Timeout can later get passesd to cancel to retrieve the data and ensure the timeout doesn't fire.

This method completes in O(1) time.

Panics

This method will panic if at is before the time that this timer wheel was created.

fn poll(&mut self, at: Instant) -> Option<T>

Queries this timer to see if any timeouts are ready to fire.

This function will advance the internal wheel to the time specified by at, returning any timeout which has happened up to that point. This method should be called in a loop until it returns None to ensure that all timeouts are processed.

Panics

This method will panic if at is before the instant that this timer wheel was created.

fn next_timeout(&self) -> Option<Instant>

Returns the instant in time that corresponds to the next timeout scheduled in this wheel.

fn cancel(&mut self, timeout: &Timeout) -> Option<T>

Cancels the specified timeout.

For timeouts previously registered via insert they can be passed back to this method to cancel the associated timeout, retrieving the value inserted if the timeout has not already fired.

This method completes in O(1) time.

Panics

This method may panic if timeout wasn't created by this timer wheel.